Add test for file copy operation between internal and external storage
This test checks if a large number of file copy operation between the
different storage media (Internal storage and SD Card) works as
expected.

It uses a physical SD Card media if it's inserted, or uses virtual SD
Card if the test device doesn't have SD Card slot or not inserted.

Bug: 72360396
Test: Run instrumentation tests from com.android.documentsui.tests
Change-Id: I1f859ceb68b89746fefaec27e11e530f10c84ab5