Update patch set 2
Patch Set 2:

Hey Todd. This change performs some provider validation to ensure DocumentsUI is working with valid DocumentsProviders prior to loading roots. But it feels like DocumentsUI should be filtering earlier...like at the time we call PackageManager#queryIntentContentProviders. Is there a way with flags or the Intent arg we can filter on permissions and exported status?

Patch-set: 2