Fix some test code for O-MR1
Symptom:
Some tests are failed on O-MR1.
1. Broken pipe error occurs when dummy data is created.

2. RejectedExecutionException is thrown
   when test files are created.

Root cause:
1. CancelFromNotificationUiTest

super.setUp() method will change the storage size to 100MB.
On the other hand, this test case uses a 500MB file as a
dummy file. Therefore, the file capacity is insufficient
and an error occurs.

2. FileCopyUiTest and FileDeleteUiTest

When too many files are created by
DocumentsProviderHelper#createDocument() method,
too many root refreshes are queued in an executor.
Then, it causes to throw RejectedExcecutionException.

Solution:
1. After excuting super.setUp(), reset the storage size again to 500MB.

2. Set a flag to prevent many refreshes.

Change-Id: Ia92e37f8e8c31d810f81fee761441c88e8635f13