Update patch set 2
Patch Set 2:

Hi Steve-san, Geoff-san,
First of all, thanks for your quick feedback to this test. Please let me answer to this higher level question from Steve-san first.

> Patch Set 2:
> 
> (2 comments)
> 
> Kuramitsu-san. Thank you for the contribution.
> 
> I have one higher level question about this change. From the perspective of DocumentsUI all storage providers are the same (each one is just a DocumentProvider implementation). So as far as DocumentsUI is concerned, copying files from any one root to any other root is all the same. Q: What sets copying to/from SDcards apart from the perspective of DocumentsUI?

That's true. We have the same understanding. The existing DocumentsUI test is a unit test as DocumentsUI application using stub providers. The new test we added here is a kind of system test thorough DocumentsUI -> ExternalStorageProvider ---> Physical storage media based on the user's file copy scenario (,and have plan of adding move operation test as noted in b/72360396).

The purpose of adding this test here are:
* SD Card support is one of the important feature for our products and it's helpful to add automation test covering this scenario so that we can detect the regression problem easily.
* Since this test emulates the user's operation on DocumentsUI app, meaning there's a few dependencies on the DocumentsUI's screen parts (i.e. UI flow, label name of button), I think it's better this test to be maintained along with DocumentsUI app.

> 
> Note, we do get some new coverage for showing/hiding internal storage.
> 
> If you're interested in exercising/validating the behavior of the DocumentsProvider that exposes SD card storage, then maybe it would be better to add a test for frameworks/base/packages/ExternalStorageProvider/src/com/android/externalstorage/ExternalStorageProvider.java or frameworks/base/core/java/com/android/internal/content/FileSystemProvider.java.

We'll be able to add test here to ensure the quality of provider as unit, but we thought system test through UI -> providers -> storage media is more meaningful.

> 
> Thanks. Let me know what you think or if you have questions.

We're happy to adapt your other feedback if this test concept is accepted as DocumentsUI test. Also we're wondering if the location of this system test code to be separated with existing function tests.

Thanks.

Patch-set: 2
Reviewer: Gerrit User 1233205 <1233205@85c56323-6fa9-3386-8a01-6480fb634889>