Refactoring: Move MimeMap to libcore.content.type.
MimeMap is not specific to Networking (URLConnection.getFileNameMap()
is not the only client) so it didn't make sense for the class to live
in the old packet. The new packet mirrors frameworks' android.content.type
which holds the logic constructing the default instance.

The new packet is specific to this purpose (libcore.net did not
previously exist) and makes for a better CorePlatformApi that we
are more likely to be able to support long term.

Bug: 124232146
Bug: 136256059
Test: CtsMimeMapTestCases
Test: CtsLibcoreTestCases
Change-Id: I60fceae7b0d1ed00896e8f3796616d6c89bb244f